LayoutTests/js/unmatching-argument-count-expected.txt

PASS eval("f()") is ""
PASS eval("f(1)") is "1"
PASS eval("f(1, 2)") is "1, 2"
PASS eval("f(1, 2, 3)") is "1, 2, 3"
PASS eval("f(1, 2, 3, 4)") is "1, 2, 3, 4"
PASS eval("f(1, 2, 3, 4, 5)") is "1, 2, 3, 4, 5"
PASS eval("f(1, 2, 3, 4, 5, 6)") is "1, 2, 3, 4, 5, 6"
PASS successfullyParsed is true

TEST COMPLETE
